refactor: extract SSH_AUTH_SOCK logic from Util into SshAuthSock (#1514)
Part of #1508 / #1514 — first slice of the Util grab-bag split.

The SSH_AUTH_SOCK discovery/validation code (issue #543 work) had
accreted onto the generic Util class. Move it to a dedicated
SshAuthSock module:

- Util::initialiseSshAuthSock()        -> SshAuthSock::initialise()
- Util::sshAuthSockOverrideStatus()    -> SshAuthSock::overrideStatus()
- Util::SshAuthSockOverrideStatus enum -> SshAuthSock::OverrideStatus
- anon isSshAgentReachable() helper    -> moved verbatim

Logic is unchanged (verbatim move); only the home and the symbol names
change. The OverrideStatus enum is now sized std::uint8_t (silences the
performance-enum-size nit it always drew on the old Util enum).

util.cpp drops its <sys/stat.h> include (only the moved socket check
used it); util keeps <QUrl> (isLaunchableWebUrl) and executor.h (WSL
findBinaryInPath).

Callers updated: main.cpp (now includes sshauthsock.h, drops the
now-unused util.h), configdialog.cpp (adds sshauthsock.h), and the
SSH tests in tst_util.cpp. src.pro gains the new module.

Build clean, tst_util 138/138 (SSH cases included), doxygen clean,
REUSE compliant.
