+++ src/util.cpp
@@ -28,7 +28,6 @@
 #ifdef Q_OS_WIN
 #include <windows.h>
 #else
-#include <sys/stat.h>
 #include <sys/time.h>
 #endif
 #include "qtpasssettings.h"
@@ -39,132 +38,6 @@
 
 QProcessEnvironment Util::_env;
 bool Util::_envInitialised = false;
-
-namespace {
-/**
- * @brief Verify that a candidate SSH_AUTH_SOCK socket has a live agent
- *        listening, before letting QtPass switch the environment to it.
- *
- * Why: a user may have a working external SSH agent (OpenSSH ssh-agent,
- * KeePassXC / 1Password / yubikey-agent / gnome-keyring) configured in
- * their shell, but its env didn't propagate to the GUI launcher. If we
- * blindly adopt whatever `gpgconf --list-dirs agent-ssh-socket` reports —
- * even when gpg-agent isn't running with SSH support, or has no keys —
- * we'd be silently switching them away from their real agent.
- *
- * Approach: spawn `ssh-add -l` with SSH_AUTH_SOCK pointed at the candidate
- * (via Executor's env-aware overload, so the parent env isn't disturbed).
- * Treat exit codes 0 (keys present) and 1 (agent alive but key list empty —
- * legitimate for YubiKey-backed setups that enumerate on tap) as
- * "reachable". Anything else, including ssh-add not being on PATH, means
- * we don't adopt the candidate.
- *
- * @param candidate Path to validate.
- * @return true if the socket has a reachable agent; false otherwise.
- */
-auto isSshAgentReachable(const QString &candidate) -> bool {
-  if (candidate.isEmpty()) {
-    return false;
-  }
-  // Build a minimal env that overrides SSH_AUTH_SOCK without polluting the
-  // parent process. systemEnvironment() captures whatever was set when
-  // QtPass started, then we override the one var of interest.
-  QProcessEnvironment env = QProcessEnvironment::systemEnvironment();
-  env.insert(QStringLiteral("SSH_AUTH_SOCK"), candidate);
-  QString out;
-  QString err;
-  const int exitCode =
-      Executor::executeBlocking(env.toStringList(), QStringLiteral("ssh-add"),
-                                {QStringLiteral("-l")}, &out, &err);
-  // OpenSSH ssh-add(1) exit codes: 0 = success / has keys, 1 = no
-  // identities present, 2 = couldn't open a connection. Anything else is
-  // also treated as unreachable (e.g. ssh-add not on PATH).
-  return exitCode == 0 || exitCode == 1;
-}
-} // namespace
-
-/**
- * @brief Probe and set SSH_AUTH_SOCK if missing — see header for full rules.
- *
- * Implementation notes:
- * - Reads QtPassSettings::getSshAuthSockOverride() for the manual override
- *   path; if non-empty, uses it verbatim (no validation — explicit override).
- * - Otherwise runs `gpgconf --list-dirs agent-ssh-socket` (gpg-agent's
- *   canonical socket reporter). On macOS, additionally tries `launchctl
- *   getenv SSH_AUTH_SOCK`.
- * - Auto-probed candidates are validated via `ssh-add -l` before adoption,
- *   so users with a different external SSH agent aren't silently switched
- *   to an empty gpg-agent SSH socket.
- * - All probes go through Executor::executeBlocking with short, bounded
- *   subprocess executions; any failure is silently absorbed (the user just
- *   doesn't get the auto-fix).
- */
-void Util::initialiseSshAuthSock() {
-  // Honour any value already in the environment — terminal launches,
-  // explicit `.desktop` Exec= overrides, and parent-process exports must win.
-  if (!qgetenv("SSH_AUTH_SOCK").isEmpty()) {
-    return;
-  }
-
-  // Manual override from settings takes precedence over auto-probe.
-  const QString override = QtPassSettings::getSshAuthSockOverride();
-  if (!override.isEmpty()) {
-    qputenv("SSH_AUTH_SOCK", override.toUtf8());
-#ifdef QT_DEBUG
-    dbg() << "Util::initialiseSshAuthSock(): set from settings override:"
-          << override;
-#endif
-    return;
-  }
-
-  // Auto-probe via gpgconf (canonical for gpg-agent's SSH support).
-  QString out;
-  QString err;
-  if (Executor::executeBlocking(
-          QStringLiteral("gpgconf"),
-          {QStringLiteral("--list-dirs"), QStringLiteral("agent-ssh-socket")},
-          &out, &err) == 0) {
-    const QString socket = out.trimmed();
-    if (!socket.isEmpty() && isSshAgentReachable(socket)) {
-      qputenv("SSH_AUTH_SOCK", socket.toUtf8());
-#ifdef QT_DEBUG
-      dbg() << "Util::initialiseSshAuthSock(): set from gpgconf:" << socket;
-#endif
-      return;
-    }
-#ifdef QT_DEBUG
-    if (!socket.isEmpty()) {
-      dbg() << "Util::initialiseSshAuthSock(): gpgconf reported" << socket
-            << "but ssh-add -l rejected it; not adopting";
-    }
-#endif
-  }
-
-#ifdef Q_OS_MACOS
-  // On macOS, GUI-launched apps may have SSH_AUTH_SOCK in launchd's
-  // per-session environment but not in the inherited process env.
-  out.clear();
-  err.clear();
-  if (Executor::executeBlocking(
-          QStringLiteral("launchctl"),
-          {QStringLiteral("getenv"), QStringLiteral("SSH_AUTH_SOCK")}, &out,
-          &err) == 0) {
-    const QString socket = out.trimmed();
-    if (!socket.isEmpty() && isSshAgentReachable(socket)) {
-      qputenv("SSH_AUTH_SOCK", socket.toUtf8());
-#ifdef QT_DEBUG
-      dbg() << "Util::initialiseSshAuthSock(): set from launchctl:" << socket;
-#endif
-    }
-#ifdef QT_DEBUG
-    else if (!socket.isEmpty()) {
-      dbg() << "Util::initialiseSshAuthSock(): launchctl reported" << socket
-            << "but ssh-add -l rejected it; not adopting";
-    }
-#endif
-  }
-#endif
-}
 
 /**
  * @brief Initializes the process environment and augments PATH with
@@ -288,29 +161,6 @@ auto Util::isPathInStore(const QString &storeRoot, const QString &candidate)
   }
   return resolved == rootCanon ||
          resolved.startsWith(rootCanon + QLatin1Char('/'));
-}
-
-auto Util::sshAuthSockOverrideStatus(const QString &path)
-    -> SshAuthSockOverrideStatus {
-  const QFileInfo fi(path);
-  if (!fi.exists()) {
-    return SshAuthSockOverrideStatus::DoesNotExist;
-  }
-  if (!fi.isReadable()) {
-    return SshAuthSockOverrideStatus::NotReadable;
-  }
-#ifdef Q_OS_UNIX
-  // S_ISSOCK isn't exposed by QFileInfo; go through stat(2). On Windows
-  // ssh-agent uses a named pipe (\\.\pipe\openssh-ssh-agent.<sid>) and the
-  // socket check would always fail, so skip it there — exists + readable is
-  // the best we can do.
-  struct stat st{};
-  if (::stat(path.toLocal8Bit().constData(), &st) != 0 ||
-      !S_ISSOCK(st.st_mode)) {
-    return SshAuthSockOverrideStatus::NotUnixDomainSocket;
-  }
-#endif
-  return SshAuthSockOverrideStatus::Valid;
 }
 
 /**
